+++ internal/service/ec2/vpc_endpoint_service_data_source_test.go
@@ -15,7 +15,7 @@ import (
 	"github.com/hashicorp/terraform-provider-aws/names"
 )
 
-func TestAccVPCEndpointServiceDataSource_gateway(t *testing.T) {
+func TestAccVPCEndpointServiceDataSource_ServiceType_gateway(t *testing.T) {
 	ctx := acctest.Context(t)
 	datasourceName := "data.aws_vpc_endpoint_service.test"
 
@@ -25,7 +25,7 @@ func TestAccVPCEndpointServiceDataSource_gateway(t *testing.T) {
 		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccVPCEndpointServiceDataSourceConfig_gateway,
+				Config: testAccVPCEndpointServiceDataSourceConfig_type("dynamodb", "Gateway"),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(datasourceName, "acceptance_required", "false"),
 					acctest.MatchResourceAttrRegionalARN(datasourceName, names.AttrARN, "ec2", regexache.MustCompile(`vpc-endpoint-service/vpce-svc-.+`)),
@@ -45,7 +45,7 @@ func TestAccVPCEndpointServiceDataSource_gateway(t *testing.T) {
 	})
 }
 
-func TestAccVPCEndpointServiceDataSource_interface(t *testing.T) {
+func TestAccVPCEndpointServiceDataSource_ServiceType_interface(t *testing.T) {
 	ctx := acctest.Context(t)
 	datasourceName := "data.aws_vpc_endpoint_service.test"
 
@@ -55,7 +55,7 @@ func TestAccVPCEndpointServiceDataSource_interface(t *testing.T) {
 		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories,
 		Steps: []resource.TestStep{
 			{
-				Config: testAccVPCEndpointServiceDataSourceConfig_interface,
+				Config: testAccVPCEndpointServiceDataSourceConfig_type("ec2", "Interface"),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(datasourceName, "acceptance_required", "false"),
 					acctest.MatchResourceAttrRegionalARN(datasourceName, names.AttrARN, "ec2", regexache.MustCompile(`vpc-endpoint-service/vpce-svc-.+`)),
@@ -168,46 +168,6 @@ func TestAccVPCEndpointServiceDataSource_CustomFilter_tags(t *testing.T) {
 	})
 }
 
-func TestAccVPCEndpointServiceDataSource_ServiceType_gateway(t *testing.T) {
-	ctx := acctest.Context(t)
-	datasourceName := "data.aws_vpc_endpoint_service.test"
-
-	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:                 func() { acctest.PreCheck(ctx, t) },
-		ErrorCheck:               acctest.ErrorCheck(t, names.EC2ServiceID),
-		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccVPCEndpointServiceDataSourceConfig_type("s3", "Gateway"),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckResourceAttrRegionalReverseDNSService(datasourceName, names.AttrServiceName, "s3"),
-					resource.TestCheckResourceAttr(datasourceName, "service_type", "Gateway"),
-				),
-			},
-		},
-	})
-}
-
-func TestAccVPCEndpointServiceDataSource_ServiceType_interface(t *testing.T) {
-	ctx := acctest.Context(t)
-	datasourceName := "data.aws_vpc_endpoint_service.test"
-
-	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:                 func() { acctest.PreCheck(ctx, t) },
-		ErrorCheck:               acctest.ErrorCheck(t, names.EC2ServiceID),
-		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccVPCEndpointServiceDataSourceConfig_type("ec2", "Interface"),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckResourceAttrRegionalReverseDNSService(datasourceName, names.AttrServiceName, "ec2"),
-					resource.TestCheckResourceAttr(datasourceName, "service_type", "Interface"),
-				),
-			},
-		},
-	})
-}
-
 // testAccCheckResourceAttrRegionalReverseDNSService ensures the Terraform state exactly matches a service reverse DNS hostname with region and partition DNS suffix
 //
 // For example: com.amazonaws.us-west-2.s3
@@ -218,18 +178,6 @@ func testAccCheckResourceAttrRegionalReverseDNSService(resourceName, attributeNa
 		return resource.TestCheckResourceAttr(resourceName, attributeName, reverseDns)(s)
 	}
 }
-
-const testAccVPCEndpointServiceDataSourceConfig_gateway = `
-data "aws_vpc_endpoint_service" "test" {
-  service = "dynamodb"
-}
-`
-
-const testAccVPCEndpointServiceDataSourceConfig_interface = `
-data "aws_vpc_endpoint_service" "test" {
-  service = "ec2"
-}
-`
 
 func testAccVPCEndpointServiceDataSourceConfig_type(service string, serviceType string) string {
 	return fmt.Sprintf(`