+++ internal/service/ec2/vpc_route_table_association.go
@@ -10,9 +10,10 @@ import (
 	"strings"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/ec2"
-	"github.com/hashicorp/aws-sdk-go-base/v2/awsv1shim/v2/tfawserr"
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/ec2"
+	awstypes "github.com/aws/aws-sdk-go-v2/service/ec2/types"
+	"github.com/hashicorp/aws-sdk-go-base/v2/tfawserr"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-provider-aws/internal/conns"
@@ -61,7 +62,7 @@ func ResourceRouteTableAssociation() *schema.Resource {
 
 func resourceRouteTableAssociationCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
-	conn := meta.(*conns.AWSClient).EC2Conn(ctx)
+	conn := meta.(*conns.AWSClient).EC2Client(ctx)
 
 	routeTableID := d.Get("route_table_id").(string)
 	input := &ec2.AssociateRouteTableInput{
@@ -78,7 +79,7 @@ func resourceRouteTableAssociationCreate(ctx context.Context, d *schema.Resource
 
 	output, err := tfresource.RetryWhenAWSErrCodeEquals(ctx, ec2PropagationTimeout,
 		func() (interface{}, error) {
-			return conn.AssociateRouteTableWithContext(ctx, input)
+			return conn.AssociateRouteTable(ctx, input)
 		},
 		errCodeInvalidRouteTableIDNotFound,
 	)
@@ -87,9 +88,9 @@ func resourceRouteTableAssociationCreate(ctx context.Context, d *schema.Resource
 		return sdkdiag.AppendErrorf(diags, "creating Route Table (%s) Association: %s", routeTableID, err)
 	}
 
-	d.SetId(aws.StringValue(output.(*ec2.AssociateRouteTableOutput).AssociationId))
+	d.SetId(aws.ToString(output.(*ec2.AssociateRouteTableOutput).AssociationId))
 
-	if _, err := WaitRouteTableAssociationCreated(ctx, conn, d.Id(), d.Timeout(schema.TimeoutCreate)); err != nil {
+	if _, err := waitRouteTableAssociationCreatedV2(ctx, conn, d.Id(), d.Timeout(schema.TimeoutCreate)); err != nil {
 		return sdkdiag.AppendErrorf(diags, "waiting for Route Table Association (%s) create: %s", d.Id(), err)
 	}
 
@@ -98,10 +99,10 @@ func resourceRouteTableAssociationCreate(ctx context.Context, d *schema.Resource
 
 func resourceRouteTableAssociationRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
-	conn := meta.(*conns.AWSClient).EC2Conn(ctx)
+	conn := meta.(*conns.AWSClient).EC2Client(ctx)
 
 	outputRaw, err := tfresource.RetryWhenNewResourceNotFound(ctx, ec2PropagationTimeout, func() (interface{}, error) {
-		return FindRouteTableAssociationByID(ctx, conn, d.Id())
+		return findRouteTableAssociationByIDV2(ctx, conn, d.Id())
 	}, d.IsNewResource())
 
 	if !d.IsNewResource() && tfresource.NotFound(err) {
@@ -114,7 +115,7 @@ func resourceRouteTableAssociationRead(ctx context.Context, d *schema.ResourceDa
 		return sdkdiag.AppendErrorf(diags, "reading Route Table Association (%s): %s", d.Id(), err)
 	}
 
-	association := outputRaw.(*ec2.RouteTableAssociation)
+	association := outputRaw.(*awstypes.RouteTableAssociation)
 
 	d.Set("gateway_id", association.GatewayId)
 	d.Set("route_table_id", association.RouteTableId)
@@ -125,15 +126,15 @@ func resourceRouteTableAssociationRead(ctx context.Context, d *schema.ResourceDa
 
 func resourceRouteTableAssociationUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
-	conn := meta.(*conns.AWSClient).EC2Conn(ctx)
+	conn := meta.(*conns.AWSClient).EC2Client(ctx)
 
 	input := &ec2.ReplaceRouteTableAssociationInput{
 		AssociationId: aws.String(d.Id()),
 		RouteTableId:  aws.String(d.Get("route_table_id").(string)),
 	}
 
-	log.Printf("[DEBUG] Updating Route Table Association: %s", input)
-	output, err := conn.ReplaceRouteTableAssociationWithContext(ctx, input)
+	log.Printf("[DEBUG] Updating Route Table Association: %v", input)
+	output, err := conn.ReplaceRouteTableAssociation(ctx, input)
 
 	// This whole thing with the resource ID being changed on update seems unsustainable.
 	// Keeping it here for backwards compatibility...
@@ -150,9 +151,9 @@ func resourceRouteTableAssociationUpdate(ctx context.Context, d *schema.Resource
 	// I don't think we'll ever reach this code for a subnet/gateway route table association.
 	// It would only come in to play for a VPC main route table association.
 
-	d.SetId(aws.StringValue(output.NewAssociationId))
+	d.SetId(aws.ToString(output.NewAssociationId))
 
-	if _, err := WaitRouteTableAssociationUpdated(ctx, conn, d.Id(), d.Timeout(schema.TimeoutUpdate)); err != nil {
+	if _, err := waitRouteTableAssociationUpdatedV2(ctx, conn, d.Id(), d.Timeout(schema.TimeoutUpdate)); err != nil {
 		return sdkdiag.AppendErrorf(diags, "waiting for Route Table Association (%s) update: %s", d.Id(), err)
 	}
 
@@ -161,7 +162,7 @@ func resourceRouteTableAssociationUpdate(ctx context.Context, d *schema.Resource
 
 func resourceRouteTableAssociationDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	var diags diag.Diagnostics
-	conn := meta.(*conns.AWSClient).EC2Conn(ctx)
+	conn := meta.(*conns.AWSClient).EC2Client(ctx)
 
 	if err := routeTableAssociationDelete(ctx, conn, d.Id(), d.Timeout(schema.TimeoutDelete)); err != nil {
 		return sdkdiag.AppendFromErr(diags, err)
@@ -180,9 +181,9 @@ func resourceRouteTableAssociationImport(ctx context.Context, d *schema.Resource
 
 	log.Printf("[DEBUG] Importing route table association, target: %s, route table: %s", targetID, routeTableID)
 
-	conn := meta.(*conns.AWSClient).EC2Conn(ctx)
+	conn := meta.(*conns.AWSClient).EC2Client(ctx)
 
-	routeTable, err := FindRouteTableByID(ctx, conn, routeTableID)
+	routeTable, err := findRouteTableByIDV2(ctx, conn, routeTableID)
 
 	if err != nil {
 		return nil, err
@@ -191,16 +192,16 @@ func resourceRouteTableAssociationImport(ctx context.Context, d *schema.Resource
 	var associationID string
 
 	for _, association := range routeTable.Associations {
-		if aws.StringValue(association.SubnetId) == targetID {
+		if aws.ToString(association.SubnetId) == targetID {
 			d.Set(names.AttrSubnetID, targetID)
-			associationID = aws.StringValue(association.RouteTableAssociationId)
+			associationID = aws.ToString(association.RouteTableAssociationId)
 
 			break
 		}
 
-		if aws.StringValue(association.GatewayId) == targetID {
+		if aws.ToString(association.GatewayId) == targetID {
 			d.Set("gateway_id", targetID)
-			associationID = aws.StringValue(association.RouteTableAssociationId)
+			associationID = aws.ToString(association.RouteTableAssociationId)
 
 			break
 		}
@@ -217,9 +218,9 @@ func resourceRouteTableAssociationImport(ctx context.Context, d *schema.Resource
 }
 
 // routeTableAssociationDelete attempts to delete a route table association.
-func routeTableAssociationDelete(ctx context.Context, conn *ec2.EC2, associationID string, timeout time.Duration) error {
+func routeTableAssociationDelete(ctx context.Context, conn *ec2.Client, associationID string, timeout time.Duration) error {
 	log.Printf("[INFO] Deleting Route Table Association: %s", associationID)
-	_, err := conn.DisassociateRouteTableWithContext(ctx, &ec2.DisassociateRouteTableInput{
+	_, err := conn.DisassociateRouteTable(ctx, &ec2.DisassociateRouteTableInput{
 		AssociationId: aws.String(associationID),
 	})
 
@@ -231,7 +232,7 @@ func routeTableAssociationDelete(ctx context.Context, conn *ec2.EC2, association
 		return fmt.Errorf("deleting Route Table Association (%s): %w", associationID, err)
 	}
 
-	if _, err := WaitRouteTableAssociationDeleted(ctx, conn, associationID, timeout); err != nil {
+	if _, err := waitRouteTableAssociationDeletedV2(ctx, conn, associationID, timeout); err != nil {
 		return fmt.Errorf("deleting Route Table Association (%s): waiting for completion: %w", associationID, err)
 	}